Update basicneuralnetworkv2.py
Here's a very bare bones implementation of a simple neural network.  For you folks who have experience with other machine learning stuff like Numpty,pytorch differs in the fact that everything is a tensor,as opposed to a vector or matrix.  If you need clarification on anything in here,look it up in the pytorch docs,or ask.